refactor(mcp): split handlers.go into 4 files and deduplicate gatherFilesFromChunks
- Created handlers_commit.go (commit-specific logic)
- Created handlers_release.go (release-specific logic)
- Created handlers_json.go (all JSON formatters + gatherFilesFromChunks)
- Removed gatherFilesFromHelperChunks duplicate from helpers.go
- Reduced handlers.go from 826 to 391 lines
- All tests pass, zero behavior changes
