+++ internal/delivery/mcp/handlers.go
@@ -2,14 +2,12 @@ package mcp
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 	"strings"
 	"time"
 
 	"github.com/Alejandro-M-P/git-courer/internal/adapters/git"
-	"github.com/Alejandro-M-P/git-courer/internal/core/domain"
 	"github.com/Alejandro-M-P/git-courer/internal/workflow"
 	mcpgo "github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
@@ -324,311 +322,6 @@ func (s *Server) handleGitWriteReview(ctx context.Context, req mcpgo.CallToolReq
 	}
 }
 
-// handleCommitOperation handles commit operations using CommitService.
-func (s *Server) handleCommitOperation(_ context.Context, req mcpgo.CallToolRequest, phase string) (*mcpgo.CallToolResult, error) {
-	requireConfirmation := s.cfg.Validation.RequireConfirmation
-	preview := req.GetBool("preview", requireConfirmation)
-
-	switch phase {
-	case "start":
-		instruction := req.GetString("instruction", "")
-
-		// Check if there's an existing plan that's still valid (not rejected)
-		if s.commitConfirm.HasBlocker() {
-			if existingPlan, _ := s.commitConfirm.ReadPlan(); existingPlan != nil && existingPlan.RejectedMessage == "" {
-				// Existing plan exists and wasn't rejected - return it directly
-				log.Printf("[DEBUG] handleCommitOperation: returning existing plan with %d messages", len(existingPlan.Messages))
-				return mcpgo.NewToolResultText(commitPlanJSON(existingPlan)), nil
-			}
-		}
-
-		if preview {
-			var rejectedMessage string
-			if s.commitConfirm.HasBlocker() {
-				if existingPlan, _ := s.commitConfirm.ReadPlan(); existingPlan != nil {
-					rejectedMessage = existingPlan.RejectedMessage
-					if rejectedMessage == "" && len(existingPlan.Messages) > 0 {
-						rejectedMessage = existingPlan.Messages[0]
-					}
-				}
-			}
-			s.llm.SetRetryContext(rejectedMessage)
-
-			s.commitConfirm.RemoveBlocker()
-
-			// Block synchronously with keepalive notifications
-			keepalive := s.startKeepalive("Preparing commit", 10*time.Second)
-			messages, chunks, deleted, _, reasoning, err := s.commitSvc.PrepareCommit(instruction)
-			close(keepalive)
-
-			if err != nil {
-				if strings.Contains(err.Error(), "[SECURITY]") {
-					s.sendSecurityErrorNotification(err.Error())
-				}
-				s.sendErrorNotification("commit", "Failed to prepare commit", map[string]any{"error": err.Error()})
-				return mcpgo.NewToolResultError("Failed to prepare commit: " + err.Error()), nil
-			}
-
-			chunkFiles := make([][]string, len(chunks))
-			var files []string
-			for i, c := range chunks {
-				chunkFiles[i] = c.Files
-				files = append(files, c.Files...)
-			}
-			files = append(files, deleted...)
-
-			plan := domain.OperationPlan{
-				Operation:    "commit",
-				Messages:     messages,
-				Files:        files,
-				Chunks:       chunkFiles,
-				DeletedFiles: deleted,
-				Instruction:  instruction,
-				Reasoning:    reasoning,
-				Preview:      strings.Join(messages, "\n"),
-			}
-
-			// Send preview via notification
-			s.sendSuccessNotification("commit", "Commit plan ready for review", &workflow.Summary{
-				Operation:     "commit",
-				FilesAffected: files,
-				Impact:        "Medium",
-				SecurityCheck: "Passed",
-				Message:       "Commit plan ready for review",
-				Reasoning:     reasoning,
-				Messages:      messages,
-			})
-
-			if err := s.commitConfirm.WritePlan(plan); err != nil {
-				s.sendErrorNotification("commit", "Failed to save plan", map[string]any{"error": err.Error()})
-				return mcpgo.NewToolResultError("Failed to save plan: " + err.Error()), nil
-			}
-			if err := s.commitConfirm.CreateBlocker(); err != nil {
-				log.Printf("[DEBUG] COMMIT_START: CreateBlocker error: %v", err)
-			}
-
-			return mcpgo.NewToolResultText(commitPlanJSON(&plan)), nil
-		}
-
-		// Non-preview: execute directly with keepalive
-		keepalive := s.startKeepalive("Running commit", 10*time.Second)
-		result, err := s.commitSvc.Execute(instruction, false)
-		close(keepalive)
-		if err != nil {
-			s.sendErrorNotification("commit", "Commit execution failed", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Commit execution failed: " + err.Error()), nil
-		}
-		s.sendSuccessNotification("commit", "Commit completed successfully", nil)
-		return mcpgo.NewToolResultText(result), nil
-
-	case "apply":
-		// For commit operations, use commit service
-		if !s.commitConfirm.HasBlocker() {
-			s.sendErrorNotification("commit", "No pending commit plan", map[string]any{"hint": "Run COMMIT_START first"})
-			return mcpgo.NewToolResultError("No active commit plan. Run COMMIT_START first."), nil
-		}
-		plan, err := s.commitConfirm.ReadPlan()
-		if err != nil || plan == nil {
-			s.commitConfirm.RemoveBlocker()
-			s.sendErrorNotification("commit", "Failed to load plan", map[string]any{"error": "Plan expired or missing"})
-			return mcpgo.NewToolResultError("Failed to load plan. Run COMMIT_START again."), nil
-		}
-		keepalive := s.startKeepalive("Executing commit", 10*time.Second)
-		result, err := s.commitSvc.ExecuteFromPlan(plan.Messages, plan.Chunks, plan.DeletedFiles, plan.Instruction)
-		close(keepalive)
-		if err != nil {
-			s.sendErrorNotification("commit", "Commit execution failed", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Commit execution failed: " + err.Error()), nil
-		}
-		s.sendSuccessNotification("commit", "Commit completed successfully", &workflow.Summary{
-			Operation:     "commit",
-			FilesAffected: plan.Files,
-			Impact:        "Medium",
-			SecurityCheck: "Verified",
-			Message:       "Commit completed successfully",
-			Reasoning:     plan.Reasoning,
-			Messages:      plan.Messages,
-		})
-		return mcpgo.NewToolResultText(result), nil
-
-	case "abort":
-		err := s.commitConfirm.DeletePlan()
-		s.llm.ClearRetryContext()
-		if err != nil {
-			return mcpgo.NewToolResultError(err.Error()), nil
-		}
-		s.sendSuccessNotification("commit", "Commit plan aborted", nil)
-		return mcpgo.NewToolResultText("Commit plan aborted."), nil
-
-	case "regenerate":
-		if !s.commitConfirm.HasBlocker() {
-			s.sendErrorNotification("commit", "No pending commit plan", map[string]any{"hint": "Run COMMIT_START first"})
-			return mcpgo.NewToolResultError("No active commit plan. Run COMMIT_START first."), nil
-		}
-
-		plan, err := s.commitConfirm.ReadPlan()
-		if err != nil || plan == nil {
-			s.commitConfirm.RemoveBlocker()
-			errMsg := "plan file not found"
-			if err != nil {
-				errMsg = err.Error()
-			}
-			s.sendErrorNotification("commit", "Failed to read plan", map[string]any{"error": errMsg})
-			return mcpgo.NewToolResultError("Failed to read plan. Run COMMIT_START again."), nil
-		}
-
-		feedback := req.GetString("feedback", "")
-		if feedback == "" {
-			feedback = "Improve the commit messages"
-		}
-
-		// Convert chunk files back to DiffChunks for regeneration
-		chunks := make([]domain.DiffChunk, len(plan.Chunks))
-		for i, files := range plan.Chunks {
-			chunks[i] = domain.DiffChunk{
-				Files: files,
-				Diff:  "", // We don't have the diff stored
-			}
-		}
-
-		keepalive := s.startKeepalive("Regenerating commit messages", 10*time.Second)
-		newMessages, err := s.llm.RegenerateMessage(plan.Messages, feedback, chunks)
-		close(keepalive)
-
-		if err != nil {
-			s.sendErrorNotification("commit", "Failed to regenerate messages", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Failed to regenerate messages: " + err.Error()), nil
-		}
-
-		// Update the plan with new messages
-		plan.Messages = newMessages
-		plan.Preview = strings.Join(newMessages, "\n")
-		
-		if err := s.commitConfirm.WritePlan(*plan); err != nil {
-			s.sendErrorNotification("commit", "Failed to update plan", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Failed to update plan: " + err.Error()), nil
-		}
-
-		s.sendSuccessNotification("commit", "Commit messages regenerated", &workflow.Summary{
-			Operation:     "commit",
-			FilesAffected: plan.Files,
-			Impact:        "Medium",
-			SecurityCheck: "Verified",
-			Message:       "Commit messages regenerated based on feedback",
-			Reasoning:     plan.Reasoning,
-			Messages:      newMessages,
-		})
-		return mcpgo.NewToolResultText(commitPlanJSON(plan)), nil
-
-	default:
-		return mcpgo.NewToolResultError("Unknown commit phase: " + phase + ". Use COMMIT_START, COMMIT_APPLY, or COMMIT_ABORT"), nil
-	}
-}
-
-// handleRelease handles release operations using ReleaseService.
-func (s *Server) handleRelease(_ context.Context, req mcpgo.CallToolRequest, phase string) (*mcpgo.CallToolResult, error) {
-	switch phase {
-	case "start":
-		instruction := req.GetString("instruction", "")
-		if instruction == "" {
-			instruction = "sacar versión"
-		}
-
-		keepalive := s.startKeepalive("Preparing release", 30*time.Second)
-		s.releaseSvc.ClearPending()
-
-		intent, commits, warnings, err := s.releaseSvc.Prepare(instruction, "")
-		close(keepalive)
-		if err != nil {
-			s.sendErrorNotification("release", "Failed to prepare release", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Failed to prepare release: " + err.Error()), nil
-		}
-
-		keepalive = s.startKeepalive("Generating changelog", 30*time.Second)
-		changelog, warningsGen, err := s.releaseSvc.Generate(commits)
-		close(keepalive)
-		if err != nil {
-			s.sendErrorNotification("release", "Failed to generate changelog", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Failed to generate changelog: " + err.Error()), nil
-		}
-
-		s.releaseSvc.SaveIntent(intent)
-		s.releaseSvc.SaveChangelog(changelog)
-
-		allWarnings := append(warnings, warningsGen...)
-		s.releaseConfirm.CreateBlocker()
-
-		authenticated, _ := s.git.IsGHAuthenticated()
-		ghStatus := "authenticated"
-		if !authenticated {
-			ghStatus = "not authenticated (will create tag only)"
-		}
-
-		s.sendSuccessNotification("release", "Release plan ready for review", map[string]any{
-			"status":        "pending_approval",
-			"tag_name":      intent.TagName,
-			"version":      intent.VersionBump,
-			"changelog":     changelog,
-			"github_auth":   ghStatus,
-			"warnings":      allWarnings,
-		})
-
-		return mcpgo.NewToolResultText(releasePlanJSON(intent, changelog, allWarnings, ghStatus)), nil
-
-	case "apply":
-		if !s.releaseConfirm.HasBlocker() {
-			s.sendErrorNotification("release", "No pending release plan", map[string]any{"hint": "Run RELEASE_START first"})
-			return mcpgo.NewToolResultError("No pending release. Run RELEASE_START first."), nil
-		}
-
-		intent, err := s.releaseSvc.LoadIntent()
-		if err != nil {
-			s.releaseConfirm.RemoveBlocker()
-			s.sendErrorNotification("release", "Failed to load intent", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError("Failed to load intent: " + err.Error()), nil
-		}
-
-		changelog, err := s.releaseSvc.LoadChangelog()
-			if err != nil {
-				s.releaseConfirm.RemoveBlocker()
-				s.sendErrorNotification("release", "Failed to load changelog", map[string]any{"error": err.Error()})
-				return mcpgo.NewToolResultError("Failed to load changelog: " + err.Error()), nil
-			}
-
-			res, err := s.applyWithBackup("release", false, func() (workflow.Result, error) {
-				output, execErr := s.releaseSvc.Execute(intent, changelog)
-				if execErr != nil {
-					return workflow.Result{}, execErr
-				}
-				return workflow.Result{
-					Status: workflow.StatusCompleted,
-					Output: output,
-					Summary: &workflow.Summary{
-						Operation: "release",
-						Impact:    "High",
-						Message:   "Release created successfully",
-					},
-				}, nil
-			})
-		if err != nil {
-			s.sendErrorNotification("release", "Release execution failed", map[string]any{"error": err.Error()})
-			return mcpgo.NewToolResultError(err.Error()), nil
-		}
-
-		s.releaseSvc.ClearPending()
-		s.sendSuccessNotification("release", "Release completed", res.Summary)
-		return mcpgo.NewToolResultText(fmt.Sprintf("Release created: %s", res.Output)), nil
-
-	case "abort":
-		s.releaseConfirm.RemoveBlocker()
-		s.releaseSvc.ClearPending()
-		s.sendSuccessNotification("release", "Release cancelled", map[string]any{})
-		return mcpgo.NewToolResultText("Release cancelled"), nil
-
-	default:
-		return mcpgo.NewToolResultError("Unknown release phase: " + phase + ". Use START, APPLY, or ABORT."), nil
-	}
-}
 
 // --- Backup helpers ---
 
@@ -696,131 +389,4 @@ func extractExplicitArgs(req mcpgo.CallToolRequest) map[string]string {
 		}
 	}
 	return result
-}
-
-func processingJSON(message string) string {
-	structuredPreview := StructuredPreview{
-		Header:   "Processing operation",
-		Sections: processingSections(message),
-		Actions:  []Action{},
-	}
-	
-	resp, _ := json.Marshal(map[string]interface{}{
-		"status":            "pending_approval",
-		"show_to_user":      "IMPORTANT: Display ALL fields below to the user before asking for confirmation. Do not summarize.",
-		"preview":           message,
-		"structured_preview": structuredPreview,
-	})
-	return string(resp)
-}
-
-func readyJSON(preview string) string {
-	structuredPreview := StructuredPreview{
-		Header:   "Review operation details",
-		Sections: genericSections("Generic git operation", preview),
-		Actions:  genericActions(),
-	}
-	
-	options := make([]string, 0, len(structuredPreview.Actions))
-	for _, action := range structuredPreview.Actions {
-		options = append(options, action.Label)
-	}
-	
-	resp, _ := json.Marshal(map[string]interface{}{
-		"status":            "pending_approval",
-		"show_to_user":      "IMPORTANT: Display ALL fields below to the user before asking for confirmation. Do not summarize.",
-		"preview":           preview,
-		"options":           options,
-		"structured_preview": structuredPreview,
-	})
-	return string(resp)
-}
-
-func formatStatus(s domain.Status) string {
-	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("Branch: %s\n", s.Branch))
-	if len(s.Files) == 0 {
-		sb.WriteString("Working tree clean\n")
-		return sb.String()
-	}
-	for _, f := range s.Files {
-		sb.WriteString(fmt.Sprintf("%s%s\n", f.Status, f.Path))
-	}
-	return sb.String()
-}
-
-func releasePlanJSON(intent *domain.ReleaseIntent, changelog string, warnings []string, ghAuth string) string {
-	structuredPreview := StructuredPreview{
-		Header:   "Review release details",
-		Sections: releaseSections(intent, changelog, warnings, ghAuth),
-		Actions:  releaseActions(),
-	}
-	
-	options := make([]string, 0, len(structuredPreview.Actions))
-	for _, action := range structuredPreview.Actions {
-		options = append(options, action.Label)
-	}
-	
-	// Calculate impact based on warnings
-	impact := "Medium - Standard release operation"
-	if len(warnings) > 0 {
-		impact = fmt.Sprintf("High - %d warning(s) detected", len(warnings))
-	}
-
-	resp, _ := json.Marshal(map[string]interface{}{
-		"status":            "pending_approval",
-		"show_to_user":      "IMPORTANT: Display ALL fields below to the user before asking for confirmation. Do not summarize.",
-		"tag_name":          intent.TagName,
-		"version":           intent.VersionBump,
-		"changelog":         changelog,
-		"github_auth":       ghAuth,
-		"warnings":          warnings,
-		"impact":            impact,
-		"options":           options,
-		"structured_preview": structuredPreview,
-	})
-	return string(resp)
-}
-
-func gatherFilesFromChunks(chunks [][]string) []string {
-	seen := make(map[string]bool)
-	var files []string
-	for _, chunk := range chunks {
-		for _, f := range chunk {
-			if !seen[f] {
-				seen[f] = true
-				files = append(files, f)
-			}
-		}
-	}
-	return files
-}
-
-func tagResultJSON(op, tag string) string {
-	return fmt.Sprintf(`{"operation": "tag_%s", "tag": %q, "status": "success"}`, op, tag)
-}
-
-func commitPlanJSON(plan *domain.OperationPlan) string {
-	structuredPreview := StructuredPreview{
-		Header:   "Review commit details",
-		Sections: commitSections(plan),
-		Actions:  commitActions(),
-	}
-	
-	options := make([]string, 0, len(structuredPreview.Actions))
-	for _, action := range structuredPreview.Actions {
-		options = append(options, action.Label)
-	}
-	
-	resp, _ := json.Marshal(map[string]interface{}{
-		"status":            "pending_approval",
-		"show_to_user":      "IMPORTANT: Display ALL fields below to the user before asking for confirmation. Do not summarize.",
-		"preview":           plan.Preview,
-		"messages":          plan.Messages,
-		"files":             gatherFilesFromChunks(plan.Chunks),
-		"reasoning":         plan.Reasoning,
-		"options":           options,
-		"structured_preview": structuredPreview,
-	})
-	return string(resp)
 }
